apps/backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-multipart
python-dotenv
pymongo
google-generativeai
opencv-python-headless
mediapipe
pypdf
jinja2
xhtml2pdf
